% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_df.R
\name{plot_df}
\alias{plot_df}
\title{Plot a scatterplot of the input dataset}
\usage{
plot_df(df, x, y, color_use)
}
\arguments{
\item{df}{the data frame}

\item{x}{the column-id of x-dim}

\item{y}{the column-id of y-dim}

\item{color}{the column-id of}
}
\description{
Plot a scatterplot of the input dataset
}
\examples{
plot_df(df = iris,x = 1,y = 2,color_use = 5)
}
