% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_iris.R
\name{plot_iris}
\alias{plot_iris}
\title{Plot a scatter plot of the iris data set}
\usage{
plot_iris(x, y)
}
\arguments{
\item{x}{the column-id of x-dim}

\item{y}{the column-id of y-dim}
}
\description{
Plot a scatter plot of the iris data set
}
\examples{
plot_iris(3,4)
}
